% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codify_RA}
\alias{codify_RA}
\title{Codified features related to rheumatoid arthritis}
\format{
A data frame with 66 rows and 4 variables
}
\usage{
data(codify_RA)
}
\description{
A dataframe containing codified features related to rheumatoid disease (PheCode:714.1) generated by ONCE. The variables are as follows:
}
\details{
\itemize{
  \item Variable. codified feature id (possible categories: PheCode/RXNORM/CCS/LOINC/Other lab)
  \item Description. full feature description
  \item relatedness_to_target_disease. Cosine similarity between the feature and PheCode:714.1.
  \item high_confidence_level. Selected features with further screening strategy.
}
}
\keyword{datasets}
