% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_RA}
\alias{cov_RA}
\title{A sample covariance matrix for rheumatoid arthritis}
\format{
A numeric matrix with 53 rows and 53 columns
}
\usage{
data(cov_RA)
}
\description{
A covariance matrix of rheumatoid arthritis (RA) related feature counts (after log transformation). Must have colnames and rownames.
}
\keyword{datasets}
