% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_RA_long}
\alias{cov_RA_long}
\title{A long-format sample covariance table for rheumatoid arthritis}
\format{
A numeric matrix with 1681 rows and 3 columns
}
\usage{
data(cov_RA_long)
}
\description{
A long-format covariance data frame of rheumatoid arthritis (RA) related feature counts (after log transformation). Must have three columns with the first two indicating the node name and the third one indicating covariance value.
}
\keyword{datasets}
