% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cui_RA}
\alias{cui_RA}
\title{NLP features related to rheumatoid arthritis}
\format{
A data frame with 897 rows and 5 variables
}
\usage{
data(cui_RA)
}
\description{
A dataframe containing NLP features related to rheumatoid disease (C0003873) generated by ONCE. The variables are as follows:
}
\details{
\itemize{
  \item cui. NLP feature id (possible categories: CUI)
  \item term. full feature description
  \item relatedness_to_target_disease. cosine similarity between the feature and C0003873
  \item importance_score. Weight combining cosine similarity, article count and term frequency that shows relatedness between features and C0003873
  \item high_confidence_level. Selected features with further screening strategy.
}
}
\keyword{datasets}
