% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.part}
\alias{dat.part}
\title{Pseudo-individual data for score prediction and model evaluation}
\format{
A data frame with 153 rows and 54 variables
}
\usage{
data(dat.part)
}
\description{
A dataframe containing pseudo patient id and codified and NLP feature counts (after log transformation). The variables are as follows:
}
\details{
\itemize{
  \item patient_num unique patient id (pseudo)
  \item utl health utility score
  \item ... Codified and NLP features
}
}
\keyword{datasets}
