% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_RA}
\alias{dict_RA}
\title{Dictionary mapping feature id to feature description}
\format{
A data frame with 183 rows and 2 variables
}
\usage{
data(dict_RA)
}
\description{
A dataframe containing id and description of features appearing in \code{input.cov}, \code{codify.feature} and \code{cuisearch.feature}. The variables are as follows:
}
\details{
\itemize{
  \item Variable. feature id (possible categories: PheCode/CCS/LOINC/RXNORM/CUI)
  \item Description. full feature description
}
}
\keyword{datasets}
