% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ehr_data}
\alias{ehr_data}
\title{Fake EHR data for preprocessing illustration (main input of the function \code{gen_cov_input})}
\format{
A data frame with 346 rows and 3 variables
}
\usage{
data(ehr_data)
}
\description{
A dataframe containing pseudo patient id, days since admission date and EHR codes. The variables are as follows:
}
\details{
\itemize{
  \item patient_num unique patient id (fake)
  \item days_since_admission time indicator
  \item code EHR code name
}
}
\keyword{datasets}
