% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_ehr}
\alias{fake_ehr}
\title{Fake EHR data for preprocessing illustration}
\format{
A data frame with 346 rows and 5 variables

A data frame with 346 rows and 5 variables
}
\usage{
data(fake_ehr)

data(fake_ehr)
}
\description{
A dataframe containing pseudo patient id and longitudinal EHR codes. The variables are as follows:

A dataframe containing pseudo patient id and longitudinal EHR codes. The variables are as follows:
}
\details{
\itemize{
  \item patient_num unique patient id (fake)
  \item days_since_admission time indicator
  \item concept_type two different EHR code types (`DIAG-ICD9` and `LAB-LOINC`) in this fake data
  \item concept_code EHR code name
  \item value lab values. -999 if the record belongs to `DIAG-ICD9`.
}

\itemize{
  \item patient_num unique patient id (fake)
  \item days_since_admission time indicator
  \item concept_type two different EHR code types (`DIAG-ICD9` and `LAB-LOINC`) in this fake data
  \item concept_code EHR code name
  \item value lab values. -999 if the record belongs to `DIAG-ICD9`.
}
}
\keyword{datasets}
