% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_ehr_label_logcount_wide}
\alias{fake_ehr_label_logcount_wide}
\title{Fake EHR data in wide format with gold labels}
\format{
A data frame with 8 rows and 50 variables
}
\usage{
data(fake_ehr_label_logcount_wide)
}
\description{
A dataframe containing pseudo patient id, gold lables and rollup counts for EHR codes. The variables are as follows:
}
\details{
\itemize{
  \item Y gold labels
  \item patient_num unique patient id (fake)
  \item ... EHR code counts
}
}
\keyword{datasets}
