% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter_df}
\alias{filter_df}
\title{Fake filters for preprocessing illustration (main input of the function \code{gen_cov_input})}
\format{
A data frame with 1 rows and 2 variables
}
\usage{
data(filter_df)
}
\description{
A dataframe containing EHR codes and the least frequency required for patient inclusion. The variables are as follows:
}
\details{
\itemize{
  \item code EHR code name
  \item freq The least frequency that is required for patients to be included to calculate the covariance matrix
}
}
\keyword{datasets}
