% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gold_label}
\alias{gold_label}
\title{Pseudo-individual data with gold label for model evaluation}
\format{
A data frame with 153 rows and 3 variables
}
\usage{
data(gold_label)
}
\description{
A dataframe containing pseudo patient id and pseudo gold label (possibly also sample probability). The variables are as follows:
}
\details{
\itemize{
  \item patient_num unique patient id (pseudo)
  \item pi sample probability (can be \code{NULL} if uniformly sampled)
  \item Y gold label (pseudo)
}
}
\keyword{datasets}
