% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{KOMAP}
\alias{KOMAP}
\title{Run komap main function}
\usage{
KOMAP(
  input.cov,
  target.code,
  target.cui,
  nm.utl,
  nm.multi = NULL,
  dict = NULL,
  codify.feature = NULL,
  cuisearch.feature = NULL,
  pred = FALSE,
  eval.real = FALSE,
  eval.sim = TRUE,
  mu0 = NULL,
  mu1 = NULL,
  var0 = NULL,
  var1 = NULL,
  prev_Y = NULL,
  B = 10000,
  dat.part = NULL,
  nm.id = NULL,
  gold.label = NULL,
  nm.pi = NULL,
  nm.y = NULL
)
}
\arguments{
\item{input.cov}{A covariance matrix of the target disease with colnames and rownames equal to variable names.}

\item{target.code}{Main phecode of the target disease.}

\item{target.cui}{Main NLP feature (CUI) of the target disease, can be NULL if NLP features are not available.}

\item{nm.utl}{Variable name of health utility score.}

\item{nm.multi}{Variable name indicating the site id. Default is NULL for the single site case.}

\item{dict}{Dictionary mapping feature id (the first column) to full description (the second column). Default is set to be NULL.}

\item{codify.feature}{A vector of codified features related to the target code. Can be generated by KESER or ONCE (Codisearch). If \code{NULL}, then use all features in \code{input.cov}.}

\item{cuisearch.feature}{A vector of NLP features related to the target code. Can be generated by ONCE (CUIsearch). If \code{NULL}, then use all features in \code{input.cov}.}

\item{pred}{If \code{TRUE}, predict disease scores for individual data.}

\item{eval.real}{If \code{TRUE}, evaluate model performance based on prediction of individual data with gold labels.}

\item{eval.sim}{If \code{TRUE}, evaluate model performance based on conditional covariance matrices, conditional mean vectors and prevalence of the target disease.}

\item{mu0}{Conditional mean vector of features given a negative label. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{mu1}{Conditional mean vector of features given a positive label. Must be provided if \code{eval.sim}=\code{TRUE} and must have the same length and in the same feature order as \code{mu0}.}

\item{var0}{Conditional covariance matrix of features given a negative label. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{var1}{Conditional covariance matrix of features given a positive label. Must be provided if \code{eval.sim}=\code{TRUE} and must have the same dimension and in the same feature order as \code{var0}.}

\item{prev_Y}{Prevalence of the target disease. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{B}{The number of samples to calculate simulated AUC, given the conditional Gaussian assumption. Default is 10000.}

\item{dat.part}{Patient-level data for prediction. Must be provided if \code{pred}=\code{TRUE} or \code{eval.real}.}

\item{nm.id}{Name of the patient id column in \code{dat.part} and \code{gold.label}. Must be provided if \code{pred}=\code{TRUE} or \code{eval.real}.}

\item{gold.label}{Gold label for patients in \code{dat.part}, containing at least one id column and one label column. Must be provided if \code{eval.real}=\code{TRUE}.}

\item{nm.pi}{The column name of sample probability in \code{gold.label} if individuals in \code{gold.label} are generated with different probabilities. Default is \code{NULL}, meaning a uniform sample distribution.}

\item{nm.y}{The column name of labels in \code{gold.label}. Must be provided if \code{eval.real}=\code{TRUE}.}
}
\value{
A list containing estimation and/or prediction and/or evaluation results by running KOMAP.
}
\description{
Run komap main function
}
\examples{
codify.feature <- codify_RA$Variable[codify_RA$select == 1]
cuisearch.feature <- cui_RA$cui[cui_RA$select == 1]
input.cov <- cov_RA
target.code <- 'PheCode:714.1'
target.cui <- 'C0003873'
nm.utl <- 'utl'
nm.pi <- 'pi'
nm.id <- 'patient_num'
nm.y <- 'Y'
out <- KOMAP(input.cov, target.code, target.cui, nm.utl, dict_RA,
                   codify.feature, cuisearch.feature,
                   pred = FALSE, eval.real = FALSE, eval.sim = TRUE,
                   mu0, mu1, var0, var1, prev_Y, B = 10000,
                   dat_part, nm.id, gold_label, nm.pi, nm.y)
}
