% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_corrupt_multi_target.R
\name{KOMAP_corrupt_multi_target}
\alias{KOMAP_corrupt_multi_target}
\title{Run komap main function}
\usage{
KOMAP_corrupt_multi_target(
  input.cov.train,
  input.cov.valid,
  is.wide = TRUE,
  targets,
  related.feature,
  nm.disease,
  nm.utl,
  nm.corrupts,
  dict = NULL,
  pred = FALSE,
  eval.real = FALSE,
  eval.sim = TRUE,
  mu0 = NULL,
  mu1 = NULL,
  var0 = NULL,
  var1 = NULL,
  prev_Y = NULL,
  B = 10000,
  dat.part = NULL,
  nm.id = NULL,
  nm.pi = NULL,
  nm.y = NULL
)
}
\arguments{
\item{is.wide}{If \code{TRUE}, \code{input.cov} is treated as a normal p by p covariance matrix. If \code{FALSE}, treat \code{input.cov} as in a long format with three columns (the first two indicating row and column names and the last one indicating covariance value). If only inputing the lower/upper triangle part, the function will automatically fill the other side assuming symmetric covariance matrix.}

\item{targets}{Main surrogates of the target disease.}

\item{related.feature}{A vector of features related to the target code. Can be generated by KESER or ONCE. If \code{NULL}, then use all features in \code{input.cov}.}

\item{nm.utl}{Variable name of health utility score.}

\item{dict}{Dictionary mapping feature id (the first column) to full description (the second column). Default is set to be NULL.}

\item{pred}{If \code{TRUE}, predict disease scores for individual data.}

\item{eval.real}{If \code{TRUE}, evaluate model performance based on prediction of individual data with gold labels.}

\item{eval.sim}{If \code{TRUE}, evaluate model performance based on conditional covariance matrices, conditional mean vectors and prevalence of the target disease.}

\item{mu0}{Conditional mean vector of features given a negative label. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{mu1}{Conditional mean vector of features given a positive label. Must be provided if \code{eval.sim}=\code{TRUE} and must have the same length and in the same feature order as \code{mu0}.}

\item{var0}{Conditional covariance matrix of features given a negative label. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{var1}{Conditional covariance matrix of features given a positive label. Must be provided if \code{eval.sim}=\code{TRUE} and must have the same dimension and in the same feature order as \code{var0}.}

\item{prev_Y}{Prevalence of the target disease. Must be provided if \code{eval.sim}=\code{TRUE}.}

\item{B}{The number of samples to calculate simulated AUC, given the conditional Gaussian assumption. Default is 10000.}

\item{dat.part}{Patient-level data for prediction. Must be provided if \code{pred}=\code{TRUE} or \code{eval.real}.}

\item{nm.id}{Name of the patient id column in \code{dat.part} and \code{gold.label}. Must be provided if \code{pred}=\code{TRUE} or \code{eval.real}.}

\item{nm.pi}{The column name of sample probability in \code{gold.label} if individuals in \code{gold.label} are generated with different probabilities. Default is \code{NULL}, meaning a uniform sample distribution.}

\item{nm.y}{The column name of labels in \code{gold.label}. Must be provided if \code{eval.real}=\code{TRUE}.}

\item{input.cov}{A covariance matrix of the target disease with colnames and rownames equal to variable names.}
}
\value{
A list containing estimation and/or prediction and/or evaluation results by running KOMAP.
}
\description{
Run komap main function
}
