% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu0}
\alias{mu0}
\title{Conditional feature mean vector (Y=0)}
\format{
A named vector with 184 entries.
}
\usage{
data(mu0)
}
\description{
A named numeric vector containing the mean of log(feature count) given negative labels for RA disease(Y=0).
}
\keyword{datasets}
