% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu1}
\alias{mu1}
\title{Conditional feature mean vector (Y=1)}
\format{
A named vector with 184 entries.
}
\usage{
data(mu1)
}
\description{
A named numeric vector containing the mean of log(feature count) given positive labels for RA disease(Y=1).
}
\keyword{datasets}
