% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecode_map}
\alias{phecode_map}
\title{Dictionary mapping from ICD9 to Phecode}
\format{
A data frame with 48 rows and 4 variables
}
\usage{
data(phecode_map)
}
\description{
A dataframe containing ICD9 and the mapped phecodes for features appearing in \code{fake_ehr}.
}
\keyword{datasets}
