% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prev_Y}
\alias{prev_Y}
\title{Prevalence of PheCode714.1}
\format{
A number.
}
\usage{
data(prev_Y)
}
\description{
A number indicating the prevalence of RA disease in the pseudo-population.
}
\keyword{datasets}
