% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rollup_dict}
\alias{rollup_dict}
\title{Fake rollup dictionary for preprocessing illustration (main input of the function \code{gen_cov_input})}
\format{
A data frame with 48 rows and 2 variables
}
\usage{
data(rollup_dict)
}
\description{
A dataframe containing EHR codes and rollup group codes. The variables are as follows:
}
\details{
\itemize{
  \item code EHR code name
  \item group Rolled-up group code
}
}
\keyword{datasets}
