% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var0}
\alias{var0}
\title{Conditional feature covariance matrix (Y=0)}
\format{
A 184*184 covariance matrix.
}
\usage{
data(var0)
}
\description{
A numeric covariance matrix of log(feature count) given negative labels for RA disease(Y=0). Must have colnames and rownames.
}
\keyword{datasets}
