% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var1}
\alias{var1}
\title{Conditional feature covariance matrix (Y=1)}
\format{
A 184*184 covariance matrix.
}
\usage{
data(var1)
}
\description{
A numeric covariance matrix of log(feature count) given positive labels for RA disease(Y=1). Must have colnames and rownames.
}
\keyword{datasets}
