% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cui_RA}
\alias{cui_RA}
\title{NLP features related to rheumatoid arthritis}
\format{
A data frame with 1183 rows and 13 variables
}
\usage{
data(cui_RA)
}
\description{
A dataframe containing NLP features related to rheumatoid disease (C0003873) generated by ONCE. The variables are as follows:
}
\details{
\itemize{
  \item cui. NLP feature id (possible categories: CUI)
  \item term. full feature description
  \item cos_cui. cosine similarity between the feature and C0003873
  \item article_count. the number of related articles containing the feature
  \item mgb_freq term-level frequency in the MGB patient cohort
  \item Acos_Ccos_d_log_va_plus. Weight combining \code{cos_cui}, \code{article_count} and \code{mgb_freq} showing relatedness between features and C0003873
  \item select. Selected features with \code{Acos_Ccos_d_log_va_plus}> mode(Acos_Ccos_d_log_va_plus) + 2 * median absolute deviation(Acos_Ccos_d_log_va_plus)
  \item ...
}
}
\keyword{datasets}
