% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_part}
\alias{dat_part}
\title{Pseudo-individual data for score prediction and model evaluation}
\format{
A data frame with 242 rows and 187 variables
}
\usage{
data(dat_part)
}
\description{
A dataframe containing pseudo patient id and codified and NLP feature counts (after log transformation). The variables are as follows:
}
\details{
\itemize{
  \item patient_num unique patient id (pseudo)
  \item Y gold label (optional)
  \item pi sample probability (optional if uniformly sampled)
  \item utl health utility score
  \item ... Codified and NLP features
}
}
\keyword{datasets}
